backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pymysql==1.1.1
sqlalchemy==2.0.31
pycryptodome==3.20.0
cryptography==42.0.8
python-jose==3.3.0
bcrypt==3.1.7
passlib==1.7.4
xmltodict==0.13.0